🚚 move apibara client into dedicated module
