--- marketplace-indexer/src/infrastructure/apibara/error.rs
@@ -1,9 +1,0 @@
-use thiserror::Error;
-
-#[derive(Debug, Error)]
-pub enum Error {
-	#[error(transparent)]
-	Connect(#[from] tonic::transport::Error),
-	#[error(transparent)]
-	Stream(#[from] tonic::Status),
-}
